Modulo 05 #FVM - Wishlist disabled

+++ src/app/shared/property-box/property-box.component.ts
@@ -15,25 +15,33 @@ export class PropertyBoxComponent implements OnInit {
   @Input() stars: string;
   @Input() whish: string;
   @Input() property_id: string;
+  @Input() isLogged: boolean;
 
   constructor(private PropertiesService: PropertiesService) { }
 
   ngOnInit() {
   }
 
-  wishlist(status, property_id){
-    if(status == true){
-      this.PropertiesService.addToWishlist(property_id)
-        .subscribe(data => {
-          this.wishlisted = status;
-        }
-      );
-    }else{
-      this.PropertiesService.removeToWishlist(property_id)
-        .subscribe(data => {
-          this.wishlisted = status;
-        }
-      );
+  wishlist($event, status, property_id){
+    if(this.isLogged) {
+      if(status == true){
+        this.PropertiesService.addToWishlist(property_id)
+          .subscribe(data => {
+            this.wishlisted = status;
+          }
+        );
+      }else{
+        this.PropertiesService.removeToWishlist(property_id)
+          .subscribe(data => {
+            this.wishlisted = status;
+          }
+        );
+      }
+    } else {
+      console.log($event);
+      $event.preventDefault();
+      $event.stopPropagation();
+      this.openLogin.emit(true);
     }
   }
 }
